// generated/src/app/api-models/menu-token.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class MenuToken {
  menuId: string;
  languageId: number;
  menuItemName: string;
  menuText: string;
  menuItemMicrohelp: string;
  keyword: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}